chore: remove jpa cascading from CA entities
Refs: XRDDEV-2323
