+++ src/central-server/admin-service/core/src/main/java/org/niis/xroad/cs/admin/core/service/CertificationServicesServiceImpl.java
@@ -25,8 +25,11 @@
  */
 package org.niis.xroad.cs.admin.core.service;
 
+import ee.ria.xroad.common.util.CertUtils;
+
 import lombok.RequiredArgsConstructor;
 import org.niis.xroad.common.exception.NotFoundException;
+import org.niis.xroad.common.exception.ValidationFailureException;
 import org.niis.xroad.cs.admin.api.dto.ApprovedCertificationService;
 import org.niis.xroad.cs.admin.api.dto.CertificateAuthority;
 import org.niis.xroad.cs.admin.api.dto.CertificateDetails;
@@ -51,6 +54,7 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
+import java.security.cert.X509Certificate;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
@@ -59,6 +63,7 @@ import static ee.ria.xroad.common.util.CryptoUtils.DEFAULT_CERT_HASH_ALGORITHM_I
 import static ee.ria.xroad.common.util.CryptoUtils.calculateCertHexHashDelimited;
 import static java.util.stream.Collectors.toList;
 import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.CERTIFICATION_SERVICE_NOT_FOUND;
+import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.INVALID_CERTIFICATE;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.AUTHENTICATION_ONLY;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.CA_ID;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.CERTIFICATE_PROFILE_INFO;
@@ -86,11 +91,32 @@ public class CertificationServicesServiceImpl implements CertificationServicesSe
     @Override
     public CertificationService add(ApprovedCertificationService certificationService) {
         CertificateProfileInfoValidator.validate(certificationService.getCertificateProfileInfo());
-        final ApprovedCaEntity approvedCaEntity = approvedCaConverter.toEntity(certificationService);
+
+        final var approvedCaEntity = new ApprovedCaEntity();
+        approvedCaEntity.setCertProfileInfo(certificationService.getCertificateProfileInfo());
+        approvedCaEntity.setAuthenticationOnly(certificationService.getTlsAuth());
+        X509Certificate certificate = handledCertificationChainRead(certificationService.getCertificate());
+        approvedCaEntity.setName(CertUtils.getSubjectCommonName(certificate));
+
+        final var caInfo = new CaInfoEntity();
+        caInfo.setCert(certificationService.getCertificate());
+        caInfo.setValidFrom(certificate.getNotBefore().toInstant());
+        caInfo.setValidTo(certificate.getNotAfter().toInstant());
+
+        approvedCaEntity.setCaInfo(caInfoRepository.save(caInfo));
+
         final ApprovedCaEntity persistedApprovedCa = approvedCaRepository.save(approvedCaEntity);
         addAuditData(persistedApprovedCa);
 
         return approvedCaConverter.convert(persistedApprovedCa);
+    }
+
+    private X509Certificate handledCertificationChainRead(byte[] certificate) {
+        try {
+            return CertUtils.readCertificateChain(certificate)[0];
+        } catch (Exception e) {
+            throw new ValidationFailureException(INVALID_CERTIFICATE);
+        }
     }
 
     @Override
@@ -132,7 +158,8 @@ public class CertificationServicesServiceImpl implements CertificationServicesSe
         final CaInfoEntity caInfo = caInfoConverter.toCaInfo(cert);
 
         final ApprovedCaEntity approvedCa = getById(certificationServiceId);
-        approvedCa.addIntermediateCa(caInfo);
+        caInfo.setApprovedCa(approvedCa);
+
         caInfoRepository.save(caInfo);
 
         auditDataHelper.put(CA_ID, certificationServiceId);
