+++ src/central-server/admin-service/core/src/test/java/org/niis/xroad/cs/admin/core/converter/OcspResponderConverterTest.java
@@ -112,7 +112,7 @@ class OcspResponderConverterTest {
         when(caInfoEntity.getId()).thenReturn(CA_ID);
 
         var ocspInfo = new OcspInfoEntity(caInfoEntity, URL, new byte[0]);
-        ocspInfo.getCaInfo().addOcspInfos(ocspInfo);
+        ocspInfo.getCaInfo().getOcspInfos().add(ocspInfo);
 
         return ocspInfo;
     }
